% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_draft}
\alias{dat_draft}
\title{Vietnam Draft Numbers}
\format{A data frame with columns:
\describe{
 \item{day_month}{A value between 1924 and 2015.}
 \item{day_year}{Maple products expressed as syrup, total in thousands of gallons.}
 \item{month}{Gross value of maple products in thousands of Canadian dollars.}
 \item{month_name}{Postal code abbreviation for territory or province.}
 \item{n69}{1969 draft order}
 \item{n70}{1970 draft order}
 \item{n71}{1971 draft order}
 \item{n72}{1972 draft order}
}}
\source{
I can't remember. Of note is that the 1969 draft was not random.
}
\usage{
dat_draft
}
\description{
Vietnam Draft Numbers
}
\examples{
\dontrun{
 dat_draft
}
}
\keyword{datasets}
