% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpr_create_package.R
\name{dpr_create_github}
\alias{dpr_create_github}
\title{Github Repo Create}
\usage{
dpr_create_github(owner_name, repo_name, public = TRUE, group = TRUE)
}
\arguments{
\item{owner_name}{is the name of the org or user that owns the repo.}

\item{repo_name}{is the name of the created data R package and repository name.}

\item{group}{defaults to TRUE and uses the group POST command to a org if FALSE uses users POST command.}
}
\description{
Github Repo Create
}
