% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpr_create_package.R
\name{dpr_delete_github}
\alias{dpr_delete_github}
\title{Github Repo Delete}
\usage{
dpr_delete_github(owner_name, repo_name)
}
\arguments{
\item{owner_name}{is the Github group or user where the package is stored.}

\item{repo_name}{is the name of the repo to delete.}
}
\description{
Github Repo Delete
}
