% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpr_export.R
\name{dpr_export}
\alias{dpr_export}
\title{Export csv files to other file formats as defined
	@param object is a character string of the csv file (with its path) or an R data.frame.}
\usage{
dpr_export(
  object,
  export_format = NULL,
  export_folder,
  details = TRUE,
  readme = NULL
)
}
\arguments{
\item{export_format}{is a character vector of formats to convert.  Defaults to ".rds".  Potential file types are .xlsx, .rds, .sav, .dta, .sas7bdat, and .json}

\item{export_folder}{is the folder where the exported data will be placed.  In this folder the a subfolder will be created for each csv file name and
selected export_format file types will be created in the folder.}

\item{details}{If TRUE then information is printed to screen during the process.  Default is TRUE.}

\item{readme}{is the data summary that describes the data.  Accepts a text string describing the provenance.}
}
\description{
Export csv files to other file formats as defined
	@param object is a character string of the csv file (with its path) or an R data.frame.
}
\examples{
dd <- read_csv(system.file("extdata", "Draft_vietnam.csv", package = "DataPushR"))
tempf <- fs::file_temp("Draft_vietnam_-_", ext = "csv")
write_csv(dd, tempf)
dpr_export(tempf, export_format = c(".rds", ".xlsx", ".sav", ".dta", ".csv", ".json", ".sas7bdat"),
export_folder=getwd())
dpr_export(dd, export_format=c(".rds",".xlsx",".sav",".dta",".csv", ".json", ".sas7bdat"), export_folder=getwd())
}
