% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpr_create_package.R
\name{dpr_info_github}
\alias{dpr_info_github}
\title{Github Repo Information}
\usage{
dpr_info_github(package_name, name)
}
\arguments{
\item{package_name}{is the name of the created data R package.}

\item{name}{is the group or github username depanding on group setting.}
}
\description{
Github Repo Information
}
\examples{
dpr_info_github("data4legos", "byuidatascience")
}
