% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpr_create_package.R
\name{dpr_push}
\alias{dpr_push}
\title{Connect to remote, Commit and Push Latest Updates to Repo}
\usage{
dpr_push(
  folder_dir = "/Users/hathawayj/git/temp_data/Test3",
  message = "'First Push from Hathaway'",
  repo_url = NULL
)
}
\arguments{
\item{folder_dir}{is the folder on your local computer where you store your git repository}

\item{message}{is the commit message to use}

\item{repo_url}{is the https url from GitHub. If NULL remote connect call not run}
}
\description{
Connect to remote, Commit and Push Latest Updates to Repo
}
