% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpr_create_package.R
\name{dpr_readme}
\alias{dpr_readme}
\title{Create readme.md in repo}
\usage{
dpr_readme(folder_dir, package_name, github_user)
}
\arguments{
\item{folder_dir}{is the folder on your local computer where you store your git repository}

\item{package_name}{is the name of the created data R package.}

\item{github_user}{is the Github group or user where the package is stored.}
}
\description{
Create readme.md in repo
}
