% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpr_create_package.R
\name{dpr_write_script}
\alias{dpr_write_script}
\title{Write data R script}
\usage{
dpr_write_script(
  folder_dir,
  r_read = "",
  r_folder_write = "data-raw",
  r_write = "",
  append_file = FALSE
)
}
\arguments{
\item{folder_dir}{is the folder on your local computer where you store your git repository}

\item{r_read}{either a path to an R script, the clipboard if `NULL`, or an `rlang::expr` object.}

\item{r_folder_write}{a path to a folder where the R script will be written}

\item{r_write}{the name of the file to write contents of `r_read`}

\item{append_file}{Whether to append `TRUE` or overwrite `FALSE` (default) the file `r_write`.}
}
\description{
Write data R script
}
