% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpr_create_package.R
\name{dpr_create_github}
\alias{dpr_create_github}
\title{Github Repo Create}
\usage{
dpr_create_github(package_name, post_text, public = TRUE)
}
\arguments{
\item{package_name}{is the name of the created data R package.}

\item{post_text}{is the api info for the POST command '/orgs/ORGNAME/repos' or '/USER/repos'}
}
\description{
Github Repo Create
}
