% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpr_create_package.R
\name{dpr_create_github}
\alias{dpr_create_github}
\title{Github Repo Create}
\usage{
dpr_create_github(
  package_name,
  post_text = c("/orgs/ORGNAME/repos", "/user/repos")[2],
  public = TRUE
)
}
\arguments{
\item{package_name}{is the name of the created data R package.}

\item{folder_dir}{is the folder on your local computer where you store your git repository}

\item{github_user}{is the Github group or user where the package is stored.}
}
\description{
Github Repo Create
}
