% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpr_create_package.R
\name{dpr_delete_github}
\alias{dpr_delete_github}
\title{Github Repo Delete}
\usage{
dpr_delete_github(package_name, owner_name, repo_name)
}
\arguments{
\item{package_name}{is the name of the created data R package.}

\item{folder_dir}{is the folder on your local computer where you store your git repository}

\item{github_user}{is the Github group or user where the package is stored.}
}
\description{
Github Repo Delete
}
