% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpr_create_package.R
\name{dpr_info_github}
\alias{dpr_info_github}
\title{Github Repo Information}
\usage{
dpr_info_github(owner_name, repo_name)
}
\arguments{
\item{owner_name}{is the Github group or user where the package is stored.}

\item{repo_name}{is the name of the repo for information.}
}
\description{
Github Repo Information
}
\examples{
dpr_info_github("data4legos", "byuidatascience")
}
